:recycle: Show wallet address on account page
